pyproject.toml
[tool.poetry]
name = "metacognitive site"
version = "1.0"
description = "A website about metacognitive therapy"
authors = ["Ivan Zaycev"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-admin = "^1.6.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
